+++ lib/lambda-cfn.js
@@ -1,4 +1,5 @@
 var path = require('path');
+var AWS = require('aws-sdk');
 var root = require('app-root-path').path;
 
 var lambdaCfn = module.exports = embed;
@@ -12,6 +13,8 @@ lambdaCfn.policy = policy;
 lambdaCfn.streambotEnv = streambotEnv;
 lambdaCfn.cloudwatch = cloudwatch;
 lambdaCfn.snsTopic = snsTopic;
+lambdaCfn.message = message;
+lambdaCfn.splitOnComma = splitOnComma;
 
 function embed(lambdaPaths, template) {
   var parts = [];
@@ -417,4 +420,28 @@ function snsTopic(options) {
       ]
     }
   };
+}
+
+function splitOnComma (str) {
+  return str.split(/\s*,\s*/);
+}
+
+function message(msg, callback) {
+
+  if (process.env.NODE_ENV == 'test') {
+    callback(null, msg);
+  } else {
+    var sns = new AWS.SNS();
+    var params = {
+      Subject: msg.subject,
+      Message:
+        msg.summary + "\n\n" +
+        JSON.stringify(msg.event, null, 2),
+      TopicArn: process.env.LambdaCfnAlarmSNSTopic
+    };
+    sns.publish(params, function(err, data) {
+      callback(err, data);
+    });
+  }
+
 }
