--- lib/utils.js
@@ -1,3 +1,0 @@
-module.exports.splitOnComma = function(str) {
-  return str.split(/\s*,\s*/);
-};
